# storage/openapi.yml
---
openapi: 3.0.0
info:
  title: Wildfire Detection
  description: This API detects wildfires with infrared technology and analyzes reports from patrolling officers.
  contact:
    email: abalcerak@mybcit.ca
  version: 1.0.0
servers:
- url: /
tags:
- name: officers
  description: Operations available to patrolling officers
- name: sensors
  description: Operations available to detection sensors
paths:
  /reports/patrols/:
    post:
      tags:
      - officers
      summary: reports the surveyed results for the region
      description: adds a new patrol report to the system for analysis
      operationId: app.report_patrol
      requestBody:
        description: Report item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatrolReport'
      responses:
        "201":
          description: report created
        "400":
          description: the input was invalid
    get:
      tags:
        - officers
      summary: gets new officer reports
      operationId: app.get_report_patrol
      description: Gets patrol reports after a timestamp
      parameters:
        - name: timestamp
          in: query
          description: Limits the number of items on a page
          schema:
            type: string
            format: date-time
            example: 2016-08-29T09:12:33.001Z
      responses:
        '200':
          description: Successfully returned a list of patrol reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatrolReport'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /reports/infrared/:
    post:
      tags:
      - sensors
      summary: reports the infrared readings in the region
      description: adds a new sensor reading to the system for analysis
      operationId: app.report_infrared
      requestBody:
        description: Report item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SensorReport'
      responses:
        "201":
          description: report created
        "400":
          description: the input was invalid
    get:
      tags:
        - sensors
      summary: gets new sensor reports
      operationId: app.get_report_infrared
      description: Gets sensor reports after a timestamp
      parameters:
        - name: timestamp
          in: query
          description: Limits the number of items on a page
          schema:
            type: string
            format: date-time
            example: 2016-08-29T09:12:33.001Z
      responses:
        '200':
          description: Successfully returned a list of sensor reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SensorReport'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
components:
  schemas:
    PatrolReport:
      required:
      - officer_id
      - reporter
      - status_code
      - timestamp
      - writeup
      - trace_id
      type: object
      properties:
        officer_id:
          type: string
          example: "132412"
        reporter:
          type: string
          example: Johnny Patrolman
        timestamp:
          format: date-time
          example: 2013-08-15T13:15:53.002Z
        status_code:
          type: integer
          example: 0
        writeup:
          type: string
          example: All clear.
        trace_id:
          type: string
          example: 7047773b-e946-411b-92ca-94c851ce3c34

    SensorReport:
      required:
      - sensor_id
      - status_code
      - temperature
      - timestamp
      - trace_id
      type: object
      properties:
        sensor_id:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        timestamp:
          type: string
          format: date-time
          example: 2013-08-16T02:11:43.001Z
        temperature:
          type: integer
          example: 14
        status_code:
          type: integer
          example: 0
        trace_id:
          type: string
          example: 7047773b-e946-411b-92ca-94c851ce3c34